MAJOR UPDATE TO SQLITE VERSION
the SQLite version now comes closest to what I have in mind for this.
Although it doesn't have much of a "front end" it is able to do
many of the things on the wish list
( https://github.com/dylan-k/bestrew#features-wishlist )

Lots of good stuff with this version. I had to learn SQL to do it!

- adding a submission record, via a view, creates new works and venues
- changing submission status also updates work status
- able to auto-count number of times a work has been submitted
- quick view of all published/active/available works
- able to store files (as BLOB data)